+++ paddle_billing/Entities/NotificationSetting.py
@@ -3,7 +3,7 @@ from dataclasses import dataclass
 
 from paddle_billing.Entities.Entity import Entity
 from paddle_billing.Entities.EventType import EventType
-from paddle_billing.Entities.NotificationSettings import NotificationSettingType
+from paddle_billing.Entities.NotificationSettings import NotificationSettingType, NotificationSettingTrafficSource
 
 
 @dataclass
@@ -17,6 +17,7 @@ class NotificationSetting(Entity):
     include_sensitive_fields: bool
     subscribed_events: list[EventType]
     endpoint_secret_key: str
+    traffic_source: NotificationSettingTrafficSource
 
     @staticmethod
     def from_dict(data: dict) -> NotificationSetting:
@@ -30,4 +31,5 @@ class NotificationSetting(Entity):
             include_sensitive_fields=data["include_sensitive_fields"],
             subscribed_events=[EventType.from_dict(event) for event in data.get("subscribed_events", [])],
             endpoint_secret_key=data["endpoint_secret_key"],
+            traffic_source=NotificationSettingTrafficSource(data["traffic_source"]),
         )