+++ paddle_billing/Entities/NotificationSettings/NotificationSettingTrafficSource.py
@@ -1,0 +1,7 @@
+from paddle_billing.PaddleStrEnum import PaddleStrEnum, PaddleStrEnumMeta
+
+
+class NotificationSettingTrafficSource(PaddleStrEnum, metaclass=PaddleStrEnumMeta):
+    All: "NotificationSettingTrafficSource" = "all"
+    Platform: "NotificationSettingTrafficSource" = "platform"
+    Simulation: "NotificationSettingTrafficSource" = "simulation"