+++ paddle_billing/Resources/NotificationSettings/Operations/CreateNotificationSetting.py
@@ -3,7 +3,7 @@ from dataclasses import dataclass
 from paddle_billing.Operation import Operation
 from paddle_billing.Undefined import Undefined
 from paddle_billing.Entities.Events import EventTypeName
-from paddle_billing.Entities.NotificationSettings import NotificationSettingType
+from paddle_billing.Entities.NotificationSettings import NotificationSettingType, NotificationSettingTrafficSource
 
 
 @dataclass
@@ -14,3 +14,4 @@ class CreateNotificationSetting(Operation):
     type: NotificationSettingType
     include_sensitive_fields: bool
     api_version: int | Undefined = Undefined()
+    traffic_source: NotificationSettingTrafficSource | Undefined = Undefined()